EN-2590 Allow Intellij to correctly import project
When Intellij imports a project, it seems to detect the names of the
modules in at least two places. One in the build.sbt file and one from
the directory name. Originally the two places had different names
('ticker-tape' for the directory and 'tickerTape' for things in the
build.sbt file). During importing the project, Intellij would detect the
two names and create extra modules. Dependencies wouldn't import and
there would be red everywhere.